move skill name's to Summary & add link to them
